requirements.txt
openai>=1.0.0
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0
